Demo 14 measured: the affinity ratio is a monotone knob on fresh telemetry, and pays nothing at this workload
The three affinity scenarios take 100 ms / 10 ms telemetry instead of the
baseline's 1000 / 200, shared by all three so the comparison still isolates
the ratio. Measured at the baseline's staleness, every ratio from 0.5 to 2.2
herds to a load CV of about 1.2 and loses to p2c on every metric: a holder
receives a second of a root's traffic before the view that would refuse it
arrives, so the ceiling never acts. With a view fresh enough to see the load
it caps, CV goes 0.32 (p2c), 0.37 (1.05), 0.75 (2.2), and goodput falls
147,043 to 145,303 to 141,314 tok/s: a hit saves an 800-token prefill, about
28 ms, in a decode-bound run, so the cache never repays the concentration.
Golden fingerprints gain demo 14's four rows and nothing else moves; the
walkthrough narrates the measured numbers instead of its earlier guesses.

+++ crates/sim-policy/src/prefix_affinity.rs
@@ -35,7 +35,7 @@ pub struct PrefixAffinity {
 pub fn make(sc: &Scenario) -> Box<dyn RoutingPolicy> {
     Box::new(PrefixAffinity {
         max_load_ratio: sc.affinity_max_load_ratio,
-        choices: sc.affinity_fallback_choices,
+        choices: sc.affinity_fallback_choices as usize,
         mean_load: None,
     })
 }
